+++ librpc/test/client_connection_test.h
@@ -25,13 +25,12 @@ TEST_F(ClientConnectionTest, ConcurrentConnectionsTest) {
   });
 
   rpc_test_utils::wait_till_server_ready(SERVER_ADDRESS, SERVER_PORT);
-
-  std::vector<rpc_client> clients(configuration_params::MAX_CONCURRENCY);
-  for (auto& client : clients) {
+  std::vector<rpc_client> clients(static_cast<unsigned long>(configuration_params::MAX_CONCURRENCY()));
+  for (auto &client : clients) {
     client.connect(SERVER_ADDRESS, SERVER_PORT);
   }
 
-  for (auto& client : clients) {
+  for (auto &client : clients) {
     client.disconnect();
   }
